+++ python/semantic_kernel/connectors/memory/azure_cosmos_db.py
@@ -23,11 +23,12 @@ from semantic_kernel.connectors.memory.mongodb import (
     MongoDBAtlasCollection,
     MongoDBAtlasStore,
 )
-from semantic_kernel.data.const import DistanceFunction, IndexKind
-from semantic_kernel.data.definitions import FieldTypes, VectorStoreCollectionDefinition
-from semantic_kernel.data.search import KernelSearchResults
-from semantic_kernel.data.vectors import (
+from semantic_kernel.data._definitions import FieldTypes, VectorStoreCollectionDefinition
+from semantic_kernel.data._search import KernelSearchResults
+from semantic_kernel.data._vectors import (
+    DistanceFunction,
     GetFilteredRecordOptions,
+    IndexKind,
     SearchType,
     TModel,
     VectorSearch,
