+++ python/semantic_kernel/connectors/memory/mongodb.py
@@ -15,10 +15,10 @@ from pymongo.driver_info import DriverInfo
 from pymongo.operations import SearchIndexModel
 
 from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
-from semantic_kernel.data.const import DistanceFunction
-from semantic_kernel.data.definitions import VectorStoreCollectionDefinition, VectorStoreField
-from semantic_kernel.data.search import KernelSearchResults
-from semantic_kernel.data.vectors import (
+from semantic_kernel.data._definitions import VectorStoreCollectionDefinition, VectorStoreField
+from semantic_kernel.data._search import KernelSearchResults
+from semantic_kernel.data._vectors import (
+    DistanceFunction,
     GetFilteredRecordOptions,
     SearchType,
     TModel,
