+++ python/semantic_kernel/connectors/memory/sql_server.py
@@ -17,11 +17,13 @@ from azure.identity.aio import DefaultAzureCredential
 from pydantic import SecretStr, ValidationError, field_validator
 
 from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
-from semantic_kernel.data.const import DISTANCE_FUNCTION_DIRECTION_HELPER, DistanceFunction, IndexKind
-from semantic_kernel.data.definitions import VectorStoreCollectionDefinition, VectorStoreField
-from semantic_kernel.data.search import KernelSearchResults
-from semantic_kernel.data.vectors import (
+from semantic_kernel.data._definitions import VectorStoreCollectionDefinition, VectorStoreField
+from semantic_kernel.data._search import KernelSearchResults
+from semantic_kernel.data._vectors import (
+    DISTANCE_FUNCTION_DIRECTION_HELPER,
+    DistanceFunction,
     GetFilteredRecordOptions,
+    IndexKind,
     SearchType,
     VectorSearch,
     VectorSearchOptions,
