+++ python/semantic_kernel/connectors/memory/weaviate.py
@@ -20,11 +20,12 @@ from weaviate.collections.collection import CollectionAsync
 from weaviate.exceptions import WeaviateClosedClientError, WeaviateConnectionError
 
 from semantic_kernel.connectors.ai.embedding_generator_base import EmbeddingGeneratorBase
-from semantic_kernel.data.const import DistanceFunction, IndexKind
-from semantic_kernel.data.definitions import VectorStoreCollectionDefinition, VectorStoreField
-from semantic_kernel.data.search import KernelSearchResults
-from semantic_kernel.data.vectors import (
+from semantic_kernel.data._definitions import VectorStoreCollectionDefinition, VectorStoreField
+from semantic_kernel.data._search import KernelSearchResults
+from semantic_kernel.data._vectors import (
+    DistanceFunction,
     GetFilteredRecordOptions,
+    IndexKind,
     SearchType,
     TModel,
     VectorSearch,
