+++ python/semantic_kernel/data/__init__.py
@@ -1,20 +1,15 @@
 # Copyright (c) Microsoft. All rights reserved.
 
 
-from semantic_kernel.data.const import (
-    DEFAULT_DESCRIPTION,
-    DEFAULT_FUNCTION_NAME,
-    DISTANCE_FUNCTION_DIRECTION_HELPER,
-    DistanceFunction,
-    IndexKind,
-)
-from semantic_kernel.data.definitions import (
+from semantic_kernel.data._definitions import (
     FieldTypes,
     VectorStoreCollectionDefinition,
     VectorStoreField,
     vectorstoremodel,
 )
-from semantic_kernel.data.search import (
+from semantic_kernel.data._search import (
+    DEFAULT_DESCRIPTION,
+    DEFAULT_FUNCTION_NAME,
     DynamicFilterFunction,
     KernelSearchResults,
     TextSearch,
@@ -22,7 +17,15 @@ from semantic_kernel.data.search import (
     create_options,
     default_dynamic_filter_function,
 )
-from semantic_kernel.data.vectors import VectorSearch, VectorSearchResult, VectorStore, VectorStoreRecordCollection
+from semantic_kernel.data._vectors import (
+    DISTANCE_FUNCTION_DIRECTION_HELPER,
+    DistanceFunction,
+    IndexKind,
+    VectorSearch,
+    VectorSearchResult,
+    VectorStore,
+    VectorStoreRecordCollection,
+)
 
 __all__ = [
     "DEFAULT_DESCRIPTION",
