+++ python/tests/integration/memory/azure_cosmos_db/conftest.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 from pydantic import BaseModel
 from pytest import fixture
 
-from semantic_kernel.data.definitions import VectorStoreField, vectorstoremodel
+from semantic_kernel.data._definitions import VectorStoreField, vectorstoremodel
 
 
 @fixture
