+++ python/tests/integration/memory/azure_cosmos_db/test_azure_cosmos_db_no_sql.py
@@ -10,7 +10,7 @@ from azure.cosmos.aio import CosmosClient
 from azure.cosmos.partition_key import PartitionKey
 
 from semantic_kernel.connectors.memory.azure_cosmos_db import AzureCosmosDBNoSQLCompositeKey, CosmosNoSqlStore
-from semantic_kernel.data.vectors import VectorStore
+from semantic_kernel.data._vectors import VectorStore
 from semantic_kernel.exceptions.memory_connector_exceptions import MemoryConnectorException
 from tests.integration.memory.vector_store_test_base import VectorStoreTestBase
