+++ python/tests/unit/connectors/memory/test_in_memory.py
@@ -3,7 +3,7 @@
 from pytest import fixture, mark, raises
 
 from semantic_kernel.connectors.memory.in_memory import InMemoryCollection, InMemoryStore
-from semantic_kernel.data.const import DistanceFunction
+from semantic_kernel.data._vectors import DistanceFunction
 from semantic_kernel.exceptions.vector_store_exceptions import VectorStoreOperationException
