+++ python/tests/unit/connectors/memory/test_postgres_store.py
@@ -17,8 +17,8 @@ from semantic_kernel.connectors.memory.postgres import (
     PostgresSettings,
     PostgresStore,
 )
-from semantic_kernel.data.const import DistanceFunction, IndexKind
-from semantic_kernel.data.definitions import VectorStoreField, vectorstoremodel
+from semantic_kernel.data._definitions import VectorStoreField, vectorstoremodel
+from semantic_kernel.data._vectors import DistanceFunction, IndexKind
 
 
 @fixture(scope="function")
