+++ python/tests/unit/connectors/memory/test_qdrant.py
@@ -7,8 +7,8 @@ from qdrant_client.async_qdrant_client import AsyncQdrantClient
 from qdrant_client.models import Datatype, Distance, FieldCondition, MatchValue, VectorParams
 
 from semantic_kernel.connectors.memory.qdrant import QdrantCollection, QdrantStore
-from semantic_kernel.data.const import DistanceFunction
-from semantic_kernel.data.definitions import VectorStoreField
+from semantic_kernel.data._definitions import VectorStoreField
+from semantic_kernel.data._vectors import DistanceFunction
 from semantic_kernel.exceptions import (
     VectorSearchExecutionException,
     VectorStoreInitializationException,
