+++ python/tests/unit/data/test_filter.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-from semantic_kernel.data.vectors import VectorSearchOptions
+from semantic_kernel.data._vectors import VectorSearchOptions
 
 
 def test_lambda_filter():
