+++ python/tests/unit/data/test_text_search.py
@@ -9,9 +9,15 @@ from pydantic import BaseModel
 
 from semantic_kernel import Kernel
 from semantic_kernel.data import TextSearch
-from semantic_kernel.data.const import DEFAULT_DESCRIPTION, DEFAULT_FUNCTION_NAME
-from semantic_kernel.data.search import KernelSearchResults, SearchOptions, TextSearchResult, create_options
-from semantic_kernel.data.vectors import VectorSearchOptions
+from semantic_kernel.data._search import (
+    DEFAULT_DESCRIPTION,
+    DEFAULT_FUNCTION_NAME,
+    KernelSearchResults,
+    SearchOptions,
+    TextSearchResult,
+    create_options,
+)
+from semantic_kernel.data._vectors import VectorSearchOptions
 from semantic_kernel.exceptions import TextSearchException
 from semantic_kernel.functions import KernelArguments, KernelParameterMetadata
 from semantic_kernel.utils.list_handler import desync_list
