+++ python/tests/unit/data/test_vector_store_model_decorator.py
@@ -10,7 +10,7 @@ from pydantic.dataclasses import dataclass as pydantic_dataclass
 from pytest import raises
 
 from semantic_kernel.data import VectorStoreCollectionDefinition, VectorStoreField
-from semantic_kernel.data.definitions import vectorstoremodel
+from semantic_kernel.data._definitions import vectorstoremodel
 from semantic_kernel.exceptions import VectorStoreModelException
